Updated lstm.py
to run lstm.py you need to run scrape historical then data processing. It runs but gets a HUGE mse i think beucase of how the traning data is scaled compared to test data- becuase of the way the first plot looks which is just plotting the train and test data from data processing (flat for first half).

to lstm2.py you just need to run srape historical and it does the processing in the start of this file.